+++ engine/simulationcraft.hpp
@@ -1959,8 +1959,8 @@ struct sim_t : private sc_thread_t
   void      interrupt();
   void      add_relative( sim_t* cousin );
   void      remove_relative( sim_t* cousin );
-  sim_progress_t progress(std::string* phase = nullptr, ssize_t index = -1 );
-  double    progress( std::string& phase, std::string* detailed = nullptr, ssize_t index = -1 );
+  sim_progress_t progress(std::string* phase = nullptr, int index = -1 );
+  double    progress( std::string& phase, std::string* detailed = nullptr, int index = -1 );
   void      detailed_progress( std::string*, int current_iterations, int total_iterations );
   void      datacollection_begin();
   void      datacollection_end();
